Cleanup of generation
- Use if guards wherever possible
- Move handle and raises to separate function so
  to have as little logic in top-level
  convert_node function.
- Move funarg to convert_definition for same
  reason.
- Make converting node and slice slightly more
  elegant.
- Rename alias to aliases in ImportAs
