+++ src/generate/convert/definition.rs
@@ -62,6 +62,21 @@ pub fn convert_def(ast: &AST, imp: &mut Imports, state: &State) -> GenResult {
                 })
             },
         },
+        Node::FunArg { vararg, var, ty, default, .. } => Core::FunArg {
+            vararg: *vararg,
+            var: Box::from(convert_node(var, imp, state)?),
+            ty: match ty {
+                Some(ty) if state.expand_ty => match &var.node {
+                    Node::_Self => None,
+                    _ => Some(Box::from(convert_node(ty, imp, state)?)),
+                },
+                _ => None,
+            },
+            default: match default {
+                Some(default) => Some(Box::from(convert_node(default, imp, state)?)),
+                None => None,
+            },
+        },
         definition => panic!("Expected definition: {:?}.", definition),
     })
 }
